% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clip_exp_vals}
\alias{clip_exp_vals}
\title{Clip Values}
\usage{
clip_exp_vals(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Input vector \code{x} with all values between -709.78 and 709.78.
}
\description{
Clip values prior to exponentiation to avoid numeric errors.
}
\examples{
clip_exp_vals(710)
}
