% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glm_fit_risk}
\alias{glm_fit_risk}
\title{Define method for converting to glm object}
\usage{
glm_fit_risk(
  x,
  y,
  weights = rep(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep(0, nobs),
  family = stats::gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\arguments{
\item{x}{input matrix with dimension n x p, every row is an observation.}

\item{y}{numeric vector for the response variable (binomial).}

\item{weights}{an optional vector of ‘prior weights’ to be used in the fitting
process. Should be NULL or a numeric vector.}

\item{start}{starting values for the parameters in the linear predictor.}

\item{etastart}{starting values for the linear predictor.}

\item{mustart}{starting values for the vector of means.}

\item{offset}{this can be used to specify an a priori known component to be
included in the linear predictor during fitting. This should be NULL or a
numeric vector of length equal to the number of cases. One or more \code{offset}
terms can be included in the formula instead or as well, and if more than one
is specified their sum is used.}

\item{family}{a description of the error distribution and link function to be
used in the model. For glm this can be a character string naming a family
function, a family function or the result of a call to a family function. For
glm.fit only the third option is supported.}

\item{control}{a list of parameters for controlling the fitting process.}

\item{intercept}{logical. Should an intercept be included in the null model?}

\item{singular.ok}{logical; if FALSE a singular fit is an error.}
}
\value{
glm object
}
\description{
Own glm.fit.own to keep coefficients provided - modified from glm's code.
}
