% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_mod.R
\name{obj_fcn}
\alias{obj_fcn}
\title{Objective function for NLL+penalty}
\usage{
obj_fcn(X, y, gamma, beta, weights, lambda0 = 0)
}
\arguments{
\item{X}{input matrix with dimension n x p, every row is an observation}

\item{y}{numeric vector for the response variable (binomial)}

\item{gamma}{scalar to rescale betas for prediction}

\item{beta}{numeric vector with p coefficients}

\item{weights}{numeric vector of length n with weights for each
observation}

\item{lambda0}{penalty coefficient for L0 term (default 0)}
}
\value{
numeric objective function value
}
\description{
Calculates the objective function for gamma, beta (NLL+penalty)
}
