% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_mod.R
\name{par_deriv}
\alias{par_deriv}
\title{Partial derivative of the negative log-likelihood}
\usage{
par_deriv(X, y, gamma, beta, weights, j)
}
\arguments{
\item{X}{input matrix with dimension n x p, every row is an observation}

\item{y}{numeric vector for the response variable (binomial)}

\item{gamma}{scalar to rescale betas for prediction}

\item{beta}{numeric vector with p coefficients}

\item{weights}{numeric vector of length n with weights for each
observation}

\item{j}{index of beta}
}
\value{
the numeric partial derivative value
}
\description{
Calculates the partial derivative for beta_j of the objective function
}
