% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randomized_rounding}
\alias{randomized_rounding}
\title{Randomly round the initialized coefficients before coordinate descent}
\usage{
randomized_rounding(beta)
}
\arguments{
\item{beta}{Numeric vector or  logistic regression coefficients initialized
before cyclical coordinate descent in \code{risk_mod()}. The first element is the
intercept and is not modified.}
}
\value{
A numeric vector with randomized rounding (apart from the first element).
}
\description{
Round each LR coefficient based on its decimal value. The decimal is the probability
of rounding the coefficient up to the next integer
}
