% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stratify_folds}
\alias{stratify_folds}
\title{Assign stratified fold ids}
\usage{
stratify_folds(y, nfolds = 10, seed = NULL)
}
\arguments{
\item{y}{numeric vector for the response variable (binomial)}

\item{nfolds}{number of folds (default 10)}

\item{seed}{An integer that is used as argument by the set.seed() for
offsetting the random number generator. Default is to leave the random number
generator alone.}
}
\value{
vector with the same length as y
}
\description{
Returns a vector of fold ids with an equal percentage of samples for each class
}
