% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{summary.risk_mod}
\alias{summary.risk_mod}
\title{Summarize Risk Model Fits}
\usage{
\method{summary}{risk_mod}(object, ...)
}
\arguments{
\item{object}{an object of class "risk_mod", usually a result of a call to risk_mod()}
}
\value{
text with intercept, nonzero coefficients, gamma, lambda, and deviance
}
\description{
Prints text that summarizes risk_mod objects
}
