% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_mod.R
\name{update_gamma_intercept}
\alias{update_gamma_intercept}
\title{Update Gamma and Intercept}
\usage{
update_gamma_intercept(X, y, beta, weights)
}
\arguments{
\item{X}{Input covariate matrix with dimension \eqn{n \times p};
every row is an observation.}

\item{y}{Numeric vector for the (binomial) response variable.}

\item{beta}{Numeric vector with \eqn{p} coefficients.}

\item{weights}{Numeric vector of length \eqn{n} with weights for each
observation. Unless otherwise specified, default will give equal weight to
each observation.}
}
\value{
A list containing the optimal gamma (numeric) and
beta (numeric vector), with gamma and \code{beta[1]} updated.
}
\description{
Finds optimal gamma value and intercept (\code{beta[1]}).
}
