% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_mod.R
\name{bisec_search}
\alias{bisec_search}
\title{Run Bisection Search}
\usage{
bisec_search(X, y, gamma, beta, weights, j, lambda0 = 0, a = -10, b = 10)
}
\arguments{
\item{X}{Input covariate matrix with dimension \eqn{n \times p};
every row is an observation.}

\item{y}{Numeric vector for the (binomial) response variable.}

\item{gamma}{Scalar to rescale coefficients for prediction.}

\item{beta}{Numeric vector with \eqn{p} coefficients.}

\item{weights}{Numeric vector of length \eqn{n} with weights for each
observation. Unless otherwise specified, default will give equal weight to
each observation.}

\item{j}{Index of \code{beta}.}

\item{lambda0}{Penalty coefficient for L0 term (default: 0).
See \code{\link[=cv_risk_mod]{cv_risk_mod()}} for \code{lambda0} tuning.}

\item{a}{Integer lower bound for coefficients (default: -10).}

\item{b}{Integer upper bound for coefficients (default: 10).}
}
\value{
Numeric vector \code{beta} with optimal value for \code{beta[j]} updated.
}
\description{
Returns optimal value on \code{beta[j]} using bisection search. For use in each
iteration of the coordinate descent algorithm.
}
