% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{coef.risk_mod}
\alias{coef.risk_mod}
\title{Extract Model Coefficients}
\usage{
\method{coef}{risk_mod}(object)
}
\arguments{
\item{object}{an object of class "risk_mod", usually a result of a call to risk_mod()}
}
\value{
numeric vector with coefficients
}
\description{
Extracts model coefficients (both nonzero and zero) from risk_mod object
}
