% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_metrics}
\alias{get_metrics}
\title{Get metrics from beta and gamma}
\usage{
get_metrics(mod, X = NULL, y = NULL, weights = NULL)
}
\arguments{
\item{mod}{riskMod object}

\item{X}{input matrix with dimensions n x p, must match dimensions of beta
in mod (default NULL)}

\item{y}{numeric vector for the response variable (binomial) of length n,
(default NULL)}

\item{weights}{numeric vector of length n with weights for each
observation (defult NULL - will give equal weights)}
}
\value{
list with deviance (dev), accuracy (acc), sensitivity (sens), and
specificity (spec)
}
\description{
Calculates deviance, accuracy, sensitivity, and specificity
}
