% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_risk}
\alias{get_risk}
\title{Get risk from score}
\usage{
get_risk(object, score)
}
\arguments{
\item{object}{an object of class "risk_mod", usually a result of a call to
risk_mod()}

\item{score}{numeric vector with score value(s)}
}
\value{
numeric vector with the same length as \code{score}
}
\description{
Returns the risk probabilities for the provided score values
}
