% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glm_fit_risk}
\alias{glm_fit_risk}
\title{Define Method for Converting to "glm" Object}
\usage{
glm_fit_risk(
  x,
  y,
  weights = rep(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep(0, nobs),
  family = stats::gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\arguments{
\item{x}{Input covariate matrix with dimension \eqn{n \times p};
every row is an observation.}

\item{y}{Numeric vector for the (binomial) response variable.}

\item{weights}{An optional vector of ‘prior weights’ to be used in the fitting
process. Should be NULL or a numeric vector.}

\item{start}{Starting values for the parameters in the linear predictor.}

\item{etastart}{Starting values for the linear predictor.}

\item{mustart}{Starting values for the vector of means.}

\item{offset}{This can be used to specify an a priori known component to be
included in the linear predictor during fitting. This should be NULL or a
numeric vector of length equal to the number of cases. One or more \code{offset}
terms can be included in the formula instead or as well, and if more than one
is specified their sum is used.}

\item{family}{A description of the error distribution and link function to be
used in the model. For glm this can be a character string naming a family
function, a family function or the result of a call to a family function. For
glm.fit only the third option is supported.}

\item{control}{A list of parameters for controlling the fitting process.}

\item{intercept}{Logical. Should an intercept be included in the null model?}

\item{singular.ok}{Logical; if FALSE a singular fit is an error.}
}
\value{
Object of class "glm".
}
\description{
Creates the "glm_fit_risk" method that will be used to convert the risk score
model to a "glm" object. Modified from \code{\link[=glm.fit]{glm.fit()}} code.
}
