% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glm_fit_risk}
\alias{glm_fit_risk}
\title{Own glm.fit.own to keep coefficients provided - modified from glm's code}
\usage{
glm_fit_risk(
  x,
  y,
  weights = rep(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep(0, nobs),
  family = stats::gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\description{
Own glm.fit.own to keep coefficients provided - modified from glm's code
}
