% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_mod.R
\name{obj_fcn}
\alias{obj_fcn}
\title{Objective Function for NLL+Penalty}
\usage{
obj_fcn(X, y, gamma, beta, weights, lambda0 = 0)
}
\arguments{
\item{X}{Input covariate matrix with dimension \eqn{n \times p};
every row is an observation.}

\item{y}{Numeric vector for the (binomial) response variable.}

\item{gamma}{Scalar to rescale coefficients for prediction.}

\item{beta}{Numeric vector with \eqn{p} coefficients.}

\item{weights}{Numeric vector of length \eqn{n} with weights for each
observation. Unless otherwise specified, default will give equal weight to
each observation.}

\item{lambda0}{Penalty coefficient for L0 term (default: 0).
See \code{\link[=cv_risk_mod]{cv_risk_mod()}} for \code{lambda0} tuning.}
}
\value{
Numeric objective function value.
}
\description{
Calculates the objective function for gamma, beta (NLL+penalty).
}
