% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_mod.R
\name{par_deriv}
\alias{par_deriv}
\title{Partial Derivative of the Negative Log-Likelihood}
\usage{
par_deriv(X, y, gamma, beta, weights, j)
}
\arguments{
\item{X}{Input covariate matrix with dimension \eqn{n \times p};
every row is an observation.}

\item{y}{Numeric vector for the (binomial) response variable.}

\item{gamma}{Scalar to rescale coefficients for prediction.}

\item{beta}{Numeric vector with \eqn{p} coefficients.}

\item{weights}{Numeric vector of length \eqn{n} with weights for each
observation. Unless otherwise specified, default will give equal weight to
each observation.}

\item{j}{Index of \code{beta}.}
}
\value{
Numeric partial derivative value.
}
\description{
Calculates the partial derivative of the objective function for \code{beta[j]}.
}
