% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{plot.cv_risk_mod}
\alias{plot.cv_risk_mod}
\title{Plot cross-validation results}
\usage{
\method{plot}{cv_risk_mod}(x, ...)
}
\arguments{
\item{x}{an object of class "cv_risk_mod", usually a result of a call to
cv_risk_mod()}
}
\value{
ggplot object
}
\description{
Plots the mean deviance for each lambda tested during cross-validation
}
