% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{plot.risk_mod}
\alias{plot.risk_mod}
\title{Plots the logistic regression curve of the risk score model}
\usage{
\method{plot}{risk_mod}(x, score_min = NULL, score_max = NULL, ...)
}
\arguments{
\item{x}{an object of class "risk_mod", usually a result of a call to
risk_mod()}

\item{score_min}{the minimum score displayed on the x-axis. The default is the
minimum score predicted from model's training data.}

\item{score_max}{the maximum score displayed on the x-axis. The default is the
maximum score predicted from model's training data.}

\item{...}{additional arguments affecting the plot produced}
}
\value{
ggplot object
}
\description{
Line plot with scores on the x-axis and risk on the y-axis.
}
