% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{predict.risk_mod}
\alias{predict.risk_mod}
\title{Model Predictions}
\usage{
\method{predict}{risk_mod}(object, newdata = NULL, type = c("link", "response", "score"))
}
\arguments{
\item{object}{an object of class "risk_mod", usually a result of a call to
risk_mod()}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the fitted linear predictors are used.}

\item{type}{the type of prediction required. The default is on the scale of
the linear predictors; the alternative "response is on the scale of the
response variable. Thus for a risk score model, the default predictions are of
log-odds (probabilities on logit scale) and type = "response" gives the
predicted risk probabilities. The "score" option returns integer risk scores.}
}
\value{
array with predictions
}
\description{
Obtains predictions from a risk score model object.
}
