% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_mod.R
\name{risk_coord_desc}
\alias{risk_coord_desc}
\title{Run Coordinate Descent}
\usage{
risk_coord_desc(
  X,
  y,
  gamma,
  beta,
  weights,
  lambda0 = 0,
  a = -10,
  b = 10,
  max_iters = 100,
  tol = 1e-05
)
}
\arguments{
\item{X}{Input covariate matrix with dimension \eqn{n \times p};
every row is an observation.}

\item{y}{Numeric vector for the (binomial) response variable.}

\item{gamma}{Scalar to rescale coefficients for prediction.}

\item{beta}{Numeric vector with \eqn{p} coefficients.}

\item{weights}{Numeric vector of length \eqn{n} with weights for each
observation. Unless otherwise specified, default will give equal weight to
each observation.}

\item{lambda0}{Penalty coefficient for L0 term (default: 0).
See \code{\link[=cv_risk_mod]{cv_risk_mod()}} for \code{lambda0} tuning.}

\item{a}{Integer lower bound for coefficients (default: -10).}

\item{b}{Integer upper bound for coefficients (default: 10).}

\item{max_iters}{Maximum number of iterations (default: 100).}

\item{tol}{Tolerance for convergence (default: 1e-5).}
}
\value{
A list containing the optimal gamma (numeric) and
beta (numeric vector).
}
\description{
Find the optimal risk score model through a coordinate descent algorithm.
At each iteration, the algorithm updates the model intercept and scalar value
(gamma) based on results from bisection search. Coordinate descent algorithm
runs until it converges or reaches the maximum number of iterations.
}
