% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_mod.R
\name{risk_mod}
\alias{risk_mod}
\title{Risk model estimation}
\usage{
risk_mod(
  X,
  y,
  gamma = NULL,
  beta = NULL,
  weights = NULL,
  lambda0 = 0,
  a = -10,
  b = 10,
  max_iters = 100,
  tol = 1e-05
)
}
\arguments{
\item{X}{input matrix with dimension n x p, every row is an observation}

\item{y}{numeric vector for the response variable (binomial)}

\item{gamma}{starting value to rescale betas for prediction (default NULL)}

\item{beta}{starting numeric vector with p coefficients (default NULL)}

\item{weights}{numeric vector of length n with weights for each
observation (defult NULL - will give equal weights)}

\item{lambda0}{penalty coefficient for L0 term (default 0)}

\item{a}{integer lower bound for betas (default -10)}

\item{b}{integer upper bound for betas (default 10)}

\item{max_iters}{maximum number of iterations (default 100)}

\item{tol}{tolerance for convergence}
}
\value{
optimal gamma (numeric) and beta (numeric vector) and corresponding
glm object (mod) as a list
}
\description{
Returns the estimated gamma and beta for the risk score model along
with a glm object with the corresponding coefficients
}
