% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_mod.R
\name{update_gamma_intercept}
\alias{update_gamma_intercept}
\title{Update gamma and beta[1]}
\usage{
update_gamma_intercept(X, y, beta, weights)
}
\arguments{
\item{X}{input matrix with dimension n x p, every row is an observation}

\item{y}{numeric vector for the response variable (binomial)}

\item{beta}{numeric vector with p coefficients}

\item{weights}{numeric vector of length n with weights for each
observation}
}
\value{
gamma and beta (with updated beta[1] value) in a list
}
\description{
Finds optimal gamma value and intercept
}
